+++ src/components/CategoryList/index.tsx
@@ -1,6 +1,4 @@
-import styled from '@emotion/styled'
-import {Link} from 'gatsby'
-import {ReactNode} from 'react'
+import * as S from './CategoryList.style';
 
 export interface CategoryListProps {
     selectedCategory: string
@@ -9,45 +7,16 @@ export interface CategoryListProps {
     }
 }
 
-interface CategoryItemProps {
-    active: boolean
-}
-
-interface GatsbyLinkProps extends CategoryItemProps {
-    children: ReactNode
-    className?: string
-    to: string
-}
-
-const CategoryListWrapper = styled.div`
-    display: flex;
-    flex-wrap: wrap;
-    width: 768px;
-    margin: 100px auto 0;
-`
-
-const CategoryItem = styled(({active, ...props}: GatsbyLinkProps)=>
-<Link {...props}/>)`
-margin-right: 20px;
-padding: 5px 0;
-font-size: 18px;
-font-weight: ${({active})=> (active? '800' : '400')};
-cursor: pointer;
-
-&:last-of-type {
-margin-right:0;
-}
-`
 
 const CategoryList = ({selectedCategory, categoryList}:CategoryListProps)=>{
     return(
-        <CategoryListWrapper>
+        <S.CategoryListWrapper>
             {Object.entries(categoryList).map(([name,count])=>(
-                <CategoryItem to={`/#category=${name}`} active={name === selectedCategory} key={name}>
+                <S.CategoryItem key={name} to={`/#category=${name}`} active={name === selectedCategory}>
                     #{name}({count})
-                </CategoryItem>
+                </S.CategoryItem>
             ))}
-        </CategoryListWrapper>
+        </S.CategoryListWrapper>
     )
 }
